+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## v1.3
 - **[ FEATURE ]** Allow to sort results in both directions
   (`--sort ASC` or `--sort DESC`)
+- **[ FEATURE ]** Print warning when unclosed open ranges are detected
+  in records before yesterday. (As this probably always a mistake.) You
+  can disable this check with the `--nowarn` flag.
 - **[ FIX ]** Don’t demand `.klg` file extensions for bookmarks
 
 ## v1.2